#143 [feat] TypeController, TypeService, TypeRouter 생성
